Compress hash-based state so it's shorter
Changes the shape of the data serialized in the URL hash so the strings
are shorter. In the v1 serialization logic, the strings were quite long
which could make sharing and bookmarking... idk, less elegant?

Anyway, in this commit the serialized data is shortened to use
single-char key names and other techniques. Larger palettes will
probably make the data longer, but proportionally and not by too much.

Fixes BNL-88